refactor : drawing_page.dart, home_page.dart, api_test.dart를 각 pages와 tests 디렉토리에 분리
