+++ lib/pages/home_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget{
+  @override
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(title: Text("Home Page")),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pushNamed(context, '/drawing');
+          },
+          child: Text("Go to Drawing Page"),
+        ),
+      ),
+    );
+  }
+}
